+++ packages/components/src/components/List/model/pagination/BatchesController.ts
@@ -24,6 +24,10 @@ export class BatchesController<T> {
   }
 
   public hasNextBatch(): boolean {
+    const filteredRowCount = this.reactTable.getFilteredRowModel().rows.length;
+    if (filteredRowCount <= this.batchSize) {
+      return false;
+    }
     return this.reactTable.getCanNextPage();
   }